feat:新增查询每月收益时间范围

+++ backend/src/main/java/com/profit/dto/BondSellRequest.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 public class BondSellRequest {
     private String startTime;
     private String endTime;
-    private Byte type;
+    private int type;
 }
